Play Again!
There is a few things needed to make this work.

UIManager has an extra element. Drag in the GameOverPanel in its place.

Next go to hierarchy and under GameOverPanel is the Play Again Button. Under OnClick() in the Inspector, please drag in UIManager for the game over script is located in there. Using the drop down go to RestartGame(int) and in the provided field write down the Build INdex of the game scene! TThats all